+++ src/main/scala/models/Peak.scala
@@ -1,6 +1,6 @@
 package models
 
-case class PeakId(id: Long)
+case class PeakId(id: Int)
 
 case class Peak(name: String, usgsid: Int, state: String, county: String,
                 map: String, elevation: Int, location: Location)
